feat(DTFS2-7715): rename get controller file
