+++ gef-ui/server/routes/facilities/amendments/index.ts
@@ -8,7 +8,7 @@ import { getWhatNeedsToChange } from '../../../controllers/amendments/what-needs
 import { getFacilityValue } from '../../../controllers/amendments/facility-value/get-facility-value';
 import { getCoverEndDate } from '../../../controllers/amendments/cover-end-date/get-cover-end-date';
 import { PORTAL_AMENDMENT_PAGES } from '../../../constants/amendments';
-import { getDoYouHaveAFacilityEndDate } from '../../../controllers/amendments/do-you-have-a-facility-end-date/do-you-have-a-facility-end-date';
+import { getDoYouHaveAFacilityEndDate } from '../../../controllers/amendments/do-you-have-a-facility-end-date/get-do-you-have-a-facility-end-date';
 
 const { WHAT_DO_YOU_NEED_TO_CHANGE, COVER_END_DATE, FACILITY_VALUE, DO_YOU_HAVE_A_FACILITY_END_DATE } = PORTAL_AMENDMENT_PAGES;
